# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Transaction Service
  description: API documentation for Transaction Service resources
  contact:
    name: Bleu Rubin
    email: budgetanalyzer@proton.me
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "1.0"
...
servers:
- url: http://localhost:8080/api
  description: Local environment (via gateway)
- url: http://localhost:8082/transaction-service
  description: Local environment (direct)
- url: https://api.budgetanalyzer.org
  description: Production environment
...
paths:
...
  /v1/admin/transactions/search:
    post:
      tags:
      - Admin - Transactions
      summary: Search transactions
      description: Paginated search over transactions
      operationId: searchTransactions
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionFilter"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
              example:
                type: INVALID_REQUEST
                message: Bad Request
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
              example:
                type: INTERNAL_ERROR
                message: Internal Server Error
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
              example:
                type: SERVICE_UNAVAILABLE
                message: Service Unavailable
...
components:
  schemas:
    TransactionResponse:
      type: object
      description: Transaction response representing a financial transaction
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the transaction
          example: 1
        accountId:
          type: string
          description: Identifier for the account associated with the transaction
          example: checking-3223
        bankName:
          type: string
          description: Name of the bank where the transaction occurred
          example: Capital One
        date:
          type: string
          format: date
          description: Date of the transaction
          example: 2025-10-14
        currencyIsoCode:
          type: string
          description: ISO currency code for the transaction
          example: USD
        amount:
          type: number
          description: Amount of the transaction
          example: 100.5
        type:
          type: string
          description: Type of the transaction
          enum:
          - CREDIT
          - DEBIT
          example: DEBIT
        description:
          type: string
          description: Description of the transaction
          example: Grocery shopping
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the transaction was created
          example: 2025-10-14T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the transaction was last updated
          example: 2025-10-14T10:30:00Z
      required:
      - amount
      - bankName
      - createdAt
      - currencyIsoCode
      - date
      - description
      - id
      - type
      - updatedAt
    ApiErrorResponse:
      description: Standard API error response format
      properties:
        type:
          type: string
          description: Error type categorization for API responses
          enum:
          - INVALID_REQUEST
          - VALIDATION_ERROR
          - NOT_FOUND
          - APPLICATION_ERROR
          - SERVICE_UNAVAILABLE
          - INTERNAL_ERROR
          - UNAUTHORIZED
          - PERMISSION_DENIED
          example: APPLICATION_ERROR
        message:
          type: string
          description: Human-readable message describing the error
          example: "CSV format: fake-bank not supported"
        code:
          type: string
          description: Machine-readable error code for specific application errors
            (required for APPLICATION_ERROR type)
          example: CSV_PARSING_ERROR
        fieldErrors:
          type: array
          description: List of field-level validation errors (populated for VALIDATION_ERROR
            type)
          items:
            $ref: "#/components/schemas/FieldError"
      required:
      - message
      - type
    FieldError:
      description: Field-level validation error details
      properties:
        field:
          type: string
          description: Field that triggered the error
          example: email
        message:
          type: string
          description: Error message
          example: must be a valid email address
        rejectedValue:
          type: object
          description: Value that caused the error
          example: invalid@email
      required:
      - field
      - message
      - rejectedValue
...
    TransactionFilter:
      type: object
      description: Filter for querying transactions by various fields
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the transaction
          example: 1
        accountId:
          type: string
          description: Identifier for the account associated with the transaction
          example: checking-3223
        bankName:
          type: string
          description: Name of the bank where the transaction occurred
          example: Capital One
        dateFrom:
          type: string
          format: date
          description: Start date for transaction date range
          example: 2025-10-01
        dateTo:
          type: string
          format: date
          description: End date for transaction date range
          example: 2025-10-14
        currencyIsoCode:
          type: string
          description: ISO currency code for the transaction
          example: USD
        minAmount:
          type: number
          description: Minimum transaction amount
          example: 10.0
        maxAmount:
          type: number
          description: Maximum transaction amount
          example: 500.0
        type:
          type: string
          description: Type of the transaction
          enum:
          - CREDIT
          - DEBIT
          example: DEBIT
        description:
          type: string
          description: Text to match in the transaction description
          example: Grocery
        createdAfter:
          type: string
          format: date-time
          description: Start of creation timestamp range
          example: 2025-10-14T00:00:00Z
        createdBefore:
          type: string
          format: date-time
          description: End of creation timestamp range
          example: 2025-10-15T00:00:00Z
        updatedAfter:
          type: string
          format: date-time
          description: Start of last update timestamp range
          example: 2025-10-14T00:00:00Z
        updatedBefore:
          type: string
          format: date-time
          description: End of last update timestamp range
          example: 2025-10-15T00:00:00Z
...